Plant_Disease_detection/requirements.txt
Flask
torch
torchvision
numpy
Pillow